# core-backend/openapi_auth_extensions.yaml
openapi: 3.1.0
info:
  title: Tochka Rosta Core - Auth Extensions (TG/MAX/VK)
  version: 1.0.0
paths:
  /api/v1/auth/send-code:
    post:
      tags: [auth]
      summary: "Send OTP code (TG/MAX)"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendCodeRequest'
      responses:
        '200':
          description: "Code sent"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  ttl:
                    type: integer
        '422':
          $ref: '#/components/responses/HTTPValidationError'

  /api/v1/auth/verify:
    post:
      tags: [auth]
      summary: "Verify OTP code and create user+tenant"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyCodeRequest'
      responses:
        '200':
          description: AuthResponse
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid or expired OTP
        '422':
          $ref: '#/components/responses/HTTPValidationError'

  /api/v1/auth/oauth/vk:
    post:
      tags: [auth]
      summary: "VK OAuth callback (code)"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OAuthVKRequest'
      responses:
        '200':
          description: AuthResponse
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid code or VK API error
        '422':
          $ref: '#/components/responses/HTTPValidationError'

components:
  schemas:
    SendCodeRequest:
      type: object
      required: ["phone", "provider"]
      properties:
        phone:
          type: string
          description: "User phone number in E.164 format"
        provider:
          type: string
          enum: [telegram, max]
          description: "Provider: telegram or max"

    VerifyCodeRequest:
      type: object
      required: ["phone", "code", "provider"]
      properties:
        phone:
          type: string
          description: "User phone number"
        code:
          type: string
          description: "OTP code"
        provider:
          type: string
          enum: [telegram, max]
          description: "Provider: telegram or max"

    OAuthVKRequest:
      type: object
      required: ["code"]
      properties:
        code:
          type: string
          description: "VK OAuth authorization code"

    AuthResponse:
      type: object
      required: ["access_token", "refresh_token", "tenant_id"]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        tenant_id:
          type: string
        user_id:
          type: string
          nullable: true

  responses:
    HTTPValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            type: object